Fix for all REST Delete & Put and screens where it called. Plus created a pop up for New Order

+++ XamarinWMS/XamarinWMS/View/Location/ShowLocation.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Plugin.Connectivity;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -26,12 +27,22 @@ namespace XamarinWMS.View
 
         public async void OnDeleteClicked(object sender, EventArgs args)
         {
-            bool accepted = await DisplayAlert("Confirm", "Are you Sure ?", "Yes", "No");
-            if (accepted)
+            bool isConnected = CrossConnectivity.Current.IsConnected;
+
+            if (isConnected)
             {
-                App.locDatabase.DeleteLoc(mSelLoc);
+                bool accepted = await DisplayAlert("Confirm", "Are you Sure ?", "Yes", "No");
+                if (accepted)
+                {
+                    App.locDatabase.DeleteLoc(mSelLoc);
+                    await App.LocManager.DeleteTaskAsync(mSelLoc);
+                }
+                await Navigation.PushAsync(new ManageLocation());
             }
-            await Navigation.PushAsync(new ManageLocation());
+            else
+            {
+                await DisplayAlert("Error", "Cannot delete a record without connection!", "OK");
+            }
         }
     }
 }
