+++ XamarinWMS/XamarinWMS/View/Product/ShowProduct.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Plugin.Connectivity;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -26,12 +27,22 @@ namespace XamarinWMS.View.Product
 
         public async void OnDeleteClicked(object sender, EventArgs args)
         {
-            bool accepted = await DisplayAlert("Confirm", "Are you Sure ?", "Yes", "No");
-            if (accepted)
+            bool isConnected = CrossConnectivity.Current.IsConnected;
+
+            if (isConnected)
             {
-                App.prodDatabase.DeleteProduct(mSelProd);
+                bool accepted = await DisplayAlert("Confirm", "Are you Sure ?", "Yes", "No");
+                if (accepted)
+                {
+                    App.prodDatabase.DeleteProduct(mSelProd);
+                    await App.ProdManager.DeleteTaskAsync(mSelProd);
+                }
+                await Navigation.PushAsync(new ManageProduct());
             }
-            await Navigation.PushAsync(new ManageProduct());
+            else
+            {
+                await DisplayAlert("Error", "Cannot delete a record without connection!", "OK");
+            }
         }
     }
 }
